moved get_ftrack_event_mongo_info to ftrack.lib
