+++ pype/modules/ftrack/ftrack_server/lib.py
@@ -19,45 +19,13 @@ import ftrack_api._centralized_storage_scenario
 import ftrack_api.event
 from ftrack_api.logging import LazyLogMessage as L
 
-from pype.modules.ftrack.lib import (
-    SERVER_HANDLERS_DIR,
-    USER_HANDLERS_DIR
-)
-from pype.api import (
-    Logger,
-    get_system_settings,
-    get_default_components,
-    decompose_url,
-    compose_url
-)
+from pype.modules.ftrack.lib import get_ftrack_event_mongo_info
 
 from .custom_db_connector import CustomDbConnector
-
+from pype.api import Logger
 
 TOPIC_STATUS_SERVER = "pype.event.server.status"
 TOPIC_STATUS_SERVER_RESULT = "pype.event.server.status.result"
-
-
-log = Logger().get_logger("ftrack_server.lib")
-
-
-def get_ftrack_event_mongo_info():
-    database_name = (
-        os.environ.get("FTRACK_EVENTS_MONGO_DB") or "pype"
-    )
-    collection_name = (
-        os.environ.get("FTRACK_EVENTS_MONGO_COL") or "ftrack_events"
-    )
-
-    mongo_url = os.environ.get("FTRACK_EVENTS_MONGO_URL")
-    if mongo_url is not None:
-        components = decompose_url(mongo_url)
-    else:
-        components = get_default_components()
-
-    uri = compose_url(**components)
-
-    return uri, components["port"], database_name, collection_name
 
 
 def check_ftrack_url(url, log_errors=True):
