+++ pype/modules/ftrack/ftrack_server/sub_event_storer.py
@@ -8,10 +8,12 @@ import pymongo
 import ftrack_api
 from ftrack_server import FtrackServer
 from pype.modules.ftrack.ftrack_server.lib import (
-    SocketSession, StorerEventHub,
-    get_ftrack_event_mongo_info,
-    TOPIC_STATUS_SERVER, TOPIC_STATUS_SERVER_RESULT
+    SocketSession,
+    StorerEventHub,
+    TOPIC_STATUS_SERVER,
+    TOPIC_STATUS_SERVER_RESULT
 )
+from pype.modules.ftrack.lib import get_ftrack_event_mongo_info
 from pype.modules.ftrack.ftrack_server.custom_db_connector import (
     CustomDbConnector
 )
