+++ pype/modules/ftrack/lib/settings.py
@@ -1,25 +1,25 @@
 import os
 from pype.api import (
     Logger,
-    get_system_settings
+    get_system_settings,
+    get_default_components,
+    decompose_url,
+    compose_url
 )
 
-
-log = Logger().get_logger("ftrack_server.lib")
+log = Logger().get_logger(__name__)
 
 FTRACK_MODULE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 SERVER_HANDLERS_DIR = os.path.join(FTRACK_MODULE_DIR, "events")
 USER_HANDLERS_DIR = os.path.join(FTRACK_MODULE_DIR, "actions")
 
 
+def get_ftrack_settings():
+    return get_system_settings()["modules"]["Ftrack"]
+
+
 def get_ftrack_url_from_settings():
-    ftrack_url = (
-        get_system_settings()
-        ["modules"]
-        ["Ftrack"]
-        ["ftrack_server"]
-    )
-    return ftrack_url
+    return get_ftrack_settings()["ftrack_server"]
 
 
 def get_server_event_handler_paths():
@@ -34,10 +34,7 @@ def get_server_event_handler_paths():
     paths.append(SERVER_HANDLERS_DIR)
     # Add additional paths from settings
     paths.extend(
-        get_system_settings()
-        ["modules"]
-        ["Ftrack"]
-        ["ftrack_events_path"]
+        get_ftrack_settings()["ftrack_events_path"]
     )
     try:
         clockify_path = clockify_event_path()
@@ -64,10 +61,7 @@ def get_user_event_handler_paths():
     paths.append(USER_HANDLERS_DIR)
     # Add additional paths from settings
     paths.extend(
-        get_system_settings()
-        ["modules"]
-        ["Ftrack"]
-        ["ftrack_actions_path"]
+        get_ftrack_settings()["ftrack_actions_path"]
     )
 
     # Filter only existing paths
@@ -96,3 +90,20 @@ def clockify_event_path():
     from pype.modules.clockify.constants import CLOCKIFY_FTRACK_SERVER_PATH
 
     return CLOCKIFY_FTRACK_SERVER_PATH
+
+
+def get_ftrack_event_mongo_info():
+    ftrack_settings = get_ftrack_settings()
+    database_name = ftrack_settings["mongo_database_name"]
+    collection_name = ftrack_settings["mongo_collection_name"]
+
+    # TODO add possibility to set in settings and use PYPE_MONGO_URL if not set
+    mongo_url = os.environ.get("FTRACK_EVENTS_MONGO_URL")
+    if mongo_url is not None:
+        components = decompose_url(mongo_url)
+    else:
+        components = get_default_components()
+
+    uri = compose_url(**components)
+
+    return uri, components["port"], database_name, collection_name
